feat: rename app/group/[groupId] to app/group/[groupSlug]
Renames the cross-group route segment to accept a slug instead of a
numeric id, resolving it via lib/group-slug.ts and 404ing on an
unresolved/unknown slug. session/_shared.tsx moves unchanged (it never
referenced groupId). Adds notFound: vi.fn() to the global next/navigation
mock, since this introduces the first notFound() calls in components
that aren't globally mocked out.

Closes #475
