--- app/group/[groupId]/effort/page.tsx
@@ -1,10 +1,0 @@
-import PayOffPage from '@/app/(routes)/effort/page';
-
-export default async function CrossGroupEffortPage({
-	params
-}: {
-	params: Promise<{ groupId: string }>;
-}) {
-	const { groupId } = await params;
-	return <PayOffPage viewedGroupId={Number(groupId)} />;
-}
