--- app/group/[groupId]/pulli/page.tsx
@@ -1,10 +1,0 @@
-import PulliPage from '@/app/(routes)/pulli/page';
-
-export default async function CrossGroupPulliPage({
-	params
-}: {
-	params: Promise<{ groupId: string }>;
-}) {
-	const { groupId } = await params;
-	return <PulliPage viewedGroupId={Number(groupId)} />;
-}
