--- app/group/[groupId]/sessions/page.tsx
@@ -1,10 +1,0 @@
-import SessionsPage from '@/app/(routes)/sessions/page';
-
-export default async function CrossGroupSessionsPage({
-	params
-}: {
-	params: Promise<{ groupId: string }>;
-}) {
-	const { groupId } = await params;
-	return <SessionsPage viewedGroupId={Number(groupId)} />;
-}
