--- app/group/[groupId]/species/page.tsx
@@ -1,10 +1,0 @@
-import AllSpeciesPage from '@/app/(routes)/species/page';
-
-export default async function CrossGroupSpeciesPage({
-	params
-}: {
-	params: Promise<{ groupId: string }>;
-}) {
-	const { groupId } = await params;
-	return <AllSpeciesPage viewedGroupId={Number(groupId)} />;
-}
