--- app/group/[groupId]/ticks/page.tsx
@@ -1,10 +1,0 @@
-import TicksPage from '@/app/(routes)/ticks/page';
-
-export default async function CrossGroupTicksPage({
-	params
-}: {
-	params: Promise<{ groupId: string }>;
-}) {
-	const { groupId } = await params;
-	return <TicksPage viewedGroupId={Number(groupId)} />;
-}
